Move _config/config.py to utils/config.py
